requirements.txt
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0
xgboost>=2.0.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
fairlearn>=0.9.0
python-dotenv>=1.0.0
joblib>=1.3.0
setuptools>=65.0.0
groq==0.3.0
requests==2.31.0